+++ .rubocop.yml
@@ -4,3 +4,6 @@ inherit_from:
 
 AllCops:
   TargetRubyVersion: 2.7
+
+Style/ExplicitBlockArgument:
+  Enabled: false
