+++ lib/phlex/html/void_elements.rb
@@ -3,6 +3,8 @@
 module Phlex::HTML::VoidElements
 	extend Phlex::Elements
 
+	REGISTERED_ELEMENTS = Concurrent::Map.new
+
 	# @!method area(**attributes, &content)
 	# 	Outputs an <code>area</code> tag
 	# 	@return [nil]
