+++ lib/phlex/svg/standard_elements.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Phlex::SVG::StandardElements
+	extend Phlex::Elements
+
+	REGISTERED_ELEMENTS = Concurrent::Map.new
+
+	register_element :svg, tag: "svg"
+end
